moved input processing into its own file, fixed a bug where the program would crash if input was given as the last argument
